+++ src/core/graph.h
@@ -36,8 +36,8 @@ public:
     neighbors_vector get_neighborhood(unsigned int); // Return the vector of neighbors of a vertex.
     unsigned int get_degree(unsigned int); // Return the degree of a vertex.
 
-    void update_state(unsigned int, float);
-    float get_state(unsigned int);
+    void update_state(unsigned int, double);
+    double get_state(unsigned int);
     void finalize_states();
 
     unsigned int get_order(); // Return the number of vertices.
